JessamineSymbolics

Documentation for JessamineSymbolics.

Introduction

This package is an interface for using Jessamine.jl with Symbolics.jl

Note: Jessamine is experimental and under development. Expect the source code to change in unpredictable ways.

JessamineSymbolics.DOTMAP

JessamineSymbolics.EPSILON

JessamineSymbolics.compile_to_function

JessamineSymbolics.eval_time_step_symbolic

JessamineSymbolics.insert_dot

JessamineSymbolics.model_basic_symbolic_output

JessamineSymbolics.model_basic_symbolic_output

JessamineSymbolics.model_symbolic_output

JessamineSymbolics.model_symbolic_output

JessamineSymbolics.replace_near_integer

JessamineSymbolics.run_genome_symbolic

JessamineSymbolics.show_symbolic

JessamineSymbolics.symbolic_form

JessamineSymbolics.DOTMAP — Constant

DOTMAP

(private) A mapping from standard functions to the symbols for their element-wise (dot) counterparts.

JessamineSymbolics.EPSILON — Constant

ε

Special symbol used to handle problems with division by zero.

JessamineSymbolics.compile_to_function — Method

compile_to_function(g_spec::GenomeSpec, agent::Agent)

Compile an agent to a function. The returned value is a callable object that takes the x's as inputs and produces an output, using the parameter vector embedded in agent. The function is compiled within the JessamineSymbolics module and should generally be used with dot broadcasting when called on inputs that are vectors.

JessamineSymbolics.eval_time_step_symbolic — Method

eval_time_step_symbolic(g_spec, genome; output_sym, scratch_sym, parameter_sym, input_sym)

Return symbolic objects representing a single time step of genome. The result is a named tuple with fields z, t, p, x for the Symbolics objects for those variables; c and c_next for the current and future cell state in symbolic form.

JessamineSymbolics.insert_dot — Method

insert_dot(e)

(private) Recursively traverses the expression e, replacing standard functions with the symbols for their element-wise (dot) counterparts according to DOTMAP.

JessamineSymbolics.model_basic_symbolic_output — Method

model_basic_symbolic_output(g_spec, agent; kw_args...)

Build a Symbolics form for the output of the final time step of running agent's genome Then use the agent's parameter vector, and feed the symbolic output of the genome as input to model result in agent.extra to make a prediction in symbolic form.

The kw_args are eventually splatted into model_predict.

JessamineSymbolics.model_basic_symbolic_output — Method

model_basic_symbolic_output(g_spec, genome, parameter, model_result; kw_args...)

Build a Symbolics form for the output of the final time step of running genome Then use the parameter vector, the symbolic output of the genome, and feed the symbolic output of the genome as input to model result in agent.extra to make a prediction in symbolic form. No simplifications are performed.

The kw_args are eventually splatted into model_predict.

Return a named tuple with lots of useful fields.

JessamineSymbolics.model_symbolic_output — Method

model_symbolic_output(g_spec, agent; kw_args...)

Build a Symbolics form for the output of the final time step of running agent's genome Then use the agent's parameter vector, and feed the symbolic output of the genome as input to model result in agent.extra to make a prediction in symbolic form. The output tuple includes the result of various simplifications.

The kw_args are eventually splatted into model_predict.

JessamineSymbolics.model_symbolic_output — Method

model_symbolic_output(g_spec, genome, parameter, model_result; kw_args...)

Build a Symbolics form for the output of the final time step of running genome Then use the parameter vector, the symbolic output of the genome, and feed the symbolic output of the genome as input to model result in agent.extra to make a prediction in symbolic form. The output tuple includes the result of various simplifications.

The kw_args are eventually splatted into model_predict.

Return a named tuple with lots of useful fields.

JessamineSymbolics.replace_near_integer — Method

replace_near_integer(expr; tolerance=1.0e-10)

Round literal floating-point numbers within a symbolic expression. Specifically, if a number a differs round(a) by less thant tolerance, it gets replaced by round(a).

JessamineSymbolics.run_genome_symbolic — Method

run_genome_symbolic(g_spec, genome; parameter_sym=:p, input_sym=:x)

Build a symbolic form for the output of the final time step of running genome. The parameter vector and input vector are Symbolics objects of the form p[j] and x[j]. The variable names can be specified with the keyword arguments.

Return a named tuple (p, x, z) where p and x, are vectors of Symbolics objects used to represent genome parameters and inputs; and z is a vector of genome outputs in symbolic form, equivalent to run_genome_to_last.

JessamineSymbolics.show_symbolic — Method

show_symbolic(g_spec, genome; output_sym, scratch_sym, parameter_sym, input_sym)

Return a matrix with three columns. The first is row indices 1, 2, etc. The second is the initial workspace vector in symbolic form. The third is the future workspace state in symbolic form.

JessamineSymbolics.symbolic_form — Method

symbolic_form(g_spec::GenomeSpec, agent::Agent)

Produce a symbolic form for an agent. Return a named tuple with field x equal to a symbol for the input array, and field vf equal to the outputs expressed in terms of x.